moved to signature validation by service manager, updated challenge to accomodate for it, moved DLSM storage to other contract

+++ src/contracts/middleware/DataLayr/DataLayrVoteWeigher.sol
@@ -183,7 +183,7 @@ contract DataLayrVoteWeigher is IVoteWeighter, IRegistrationManager {
         latestTime = _latestTime;
     }
 
-    function getOperatorId(address operator) public view returns(uint256) {
+    function getOperatorId(address operator) public view returns(uint32) {
         return registry[operator].id;
     }
 }
